Make original exception available in CircuitOpenError
When the circuit breaker is open, the exception raised doesn't tell us
what caused the circuit breaker to trip. Since ruby 2.1, there's a
standard interface for defining a nested exception using the #cause
method [1].

This change makes the last exception that caused the circuit breaker to
fail available from the raised CircuitOpenError.

[1] http://blog.honeybadger.io/nested-errors-in-ruby-with-exception-cause/